* refactor: adapt block BlockCherryWood * refactor: adapt block BlockDaylightDetector * refactor: adapt block BlockRedNetherBrick * refactor: adapt block BlockStoneBlockSlab2 * refactor: adapt block BlockSweetBerryBush * add and adapt ObjectAzaleaTree * adapt some items to block changes * clean code:   * remove double space,   * format @NotNull   * format @Nullable   * enhanced switch

+++ src/main/java/cn/nukkit/level/generator/object/ObjectJungleBigTree.java
@@ -25,7 +25,7 @@ public class ObjectJungleBigTree extends HugeTreesGenerator {
         } else {
             this.createCrown(level, position.up(height), 2);
 
-            for (int j = (int) position.getY() + height - 2 - rand.nextInt(4); j > position.getY() + height / 2; j -= 2 + rand.nextInt(4)) {
+            for (int j = (int) position.getY() + height - 2 - rand.nextInt(4); j > position.getY() + (double) height / 2; j -= 2 + rand.nextInt(4)) {
                 float f = rand.nextFloat() * ((float) Math.PI * 2F);
                 int k = (int) (position.getX() + (0.5F + MathHelper.cos(f) * 4.0F));
                 int l = (int) (position.getZ() + (0.5F + MathHelper.sin(f) * 4.0F));
